Unsign with various fixes.
This adds support for unsigning a package.

This also fixes a number of other issues, particularly found when
removing content. The ZIP stream was not truncated before the file
was written, and if the new content was shorter than the old content,
then parts of the old content would not be removed.

This also fixes an issue where the origin relationship may not have
been removed correctly.