Add some test support for Azure Signing.
Currently this cannot run in the CI environment, though it may be
possible with a client certificate. The "threat" being that anyone could
submit a PR that will automatically run and dump the credentials. A
client certificate with a non-exportable key might suffice, especially
since this is a self-signed certificate.